Crate salvo

Re-exports

pub use salvo_core as core;

pub use salvo_macro;

Modules

depot

http
Types that map to concepts in HTTP.

logging

prelude

routing

server

writer

Structs

CatcherImpl

Depot

Error
The Error type, a wrapper around a dynamic error type.

Protocol
Protocol used to serve content.

Request
The Request given to all Middleware.

Response
The response representation given to Middleware

Router

Server
The main Novel type: used to mount routes and catchers and launch the application.

ServerConfig

Traits

Catcher

Handler

Writer