Trait salvo::http::headers::Header

pub trait Header {
    pub fn name() -> &'static HeaderName;

    pub fn decode<'i, I>(values: &mut I) -> Result<Self, Error>
    where
        I: Iterator<Item = &'i HeaderValue>;

    pub fn encode<E>(&self, values: &mut E)
    where
        E: Extend<HeaderValue>;
}

A trait for any object that will represent a header field and value.

This trait represents the construction and identification of headers, and contains trait-object unsafe methods.

Required methods

pub fn name() -> &'static HeaderName

The name of this header.

pub fn decode<'i, I>(values: &mut I) -> Result<Self, Error> where
I: Iterator<Item = &'i HeaderValue>,

Decode this type from an iterator of HeaderValues.

pub fn encode<E>(&self, values: &mut E) where
E: Extend<HeaderValue>,

Encode this type to a HeaderMap.

This function should be infallible. Any errors converting to a HeaderValue should have been caught when parsing or constructing this value.

Loading content...